Update createFurnitureDatabase.sql
Changed the schema name to inf124grp06 as using furniture will not allow us to do so...access was being denied. Also, I added AUTO_INCREMENT on all the tables... missing this entry was an error of "id" not found. That is all. Now I successfully updated the database in andromeda-6