Added files via upload
The file is for pulling corresponding city and state when a customer type in a zip code via order.html form... Please I need to some help to complete... I'm not sure so far how to pass the zip code to this php file and how to get the city and state back. I already set the query and some data... but I do no think it is enough.